# services/api/openapi.yaml
openapi: 3.1.0
info:
  title: Diorama API
  version: 0.1.0
  summary: Hosted map data, search, transit and verified drops for Diorama apps.
  description: |
    The paid hosted service next to the open-source Diorama SDK.

    ## Authentication
    Every endpoint except `/v1/health` needs an API key:
    `Authorization: Bearer <key>`. Map clients that cannot set headers may pass
    `?key=<key>` **only** on `GET /v1/worlds/*` and `GET /v1/tiles/*`; elsewhere a
    query key is rejected with `401 QUERY_KEY_NOT_ALLOWED`.

    Keys have a role: `client` (apps/devices), `server` (app backends) and
    `admin` (app owner). Keys are stored as SHA-256 hashes only.

    ## Usage and billing
    Billable requests add weighted units to the key's UTC-month counter:
    `tile` 1, `world` 20, `search` 5 (search and reverse), `transit` 2,
    `drops` 2, `collect` 10. A request is billable when its status is < 400,
    or 422 for collect verification. Free keys receive `429 QUOTA_EXCEEDED`
    once the monthly quota would be exceeded; pro keys continue and the excess
    is recorded as overage. Metered responses carry `X-Diorama-Usage`,
    `X-RateLimit-Limit` and `X-RateLimit-Remaining`.

    ## Errors
    Errors are JSON `{ "error": { "code": "...", "message": "..." } }` with a
    stable `code` (see the `ErrorCode` schema).
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:8787
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/drops/campaigns:
    post:
      tags: [Drops]
      operationId: createCampaign
      summary: Create a drop campaign
      description: |
        Requires role `admin` or `server`. Drops are not stored: every
        campaign × time window × geohash (precision 6) cell yields its drops from a
        seeded PRNG, so all clients see the same drops. Not metered.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CampaignCreate' }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                type: object
                required: [campaign]
                properties:
                  campaign: { $ref: '#/components/schemas/Campaign' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
  /v1/drops/nearby:
    get:
      tags: [Drops]
      operationId: getNearbyDrops
      summary: Drops near a position in the current window
      description: Sorted by distance, at most 1000. Billable unit `drops`.
      parameters:
        - { name: lng, in: query, required: true, schema: { type: number, minimum: -180, maximum: 180 } }
        - { name: lat, in: query, required: true, schema: { type: number, minimum: -90, maximum: 90 } }
        - { name: radius, in: query, required: false, schema: { type: number, minimum: 1, maximum: 3000, default: 500 } }
        - { name: channel, in: query, required: true, schema: { type: string, pattern: '^[A-Za-z0-9_.:-]{1,64}$' } }
      responses:
        '200':
          description: Drops.
          content:
            application/json:
              schema:
                type: object
                required: [drops, generatedAt, expiresAt]
                properties:
                  drops:
                    type: array
                    items: { $ref: '#/components/schemas/DropSpec' }
                  generatedAt: { type: integer, description: Server time (ms). }
                  expiresAt:
                    type: [integer, 'null']
                    description: End of the earliest current window (ms); refetch after it.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/QuotaExceeded' }
  /v1/drops/collect:
    post:
      tags: [Drops]
      operationId: collectDrop
      summary: Verify a drop collection
      description: |
        The SDK judges collection on-device and forwards it here. Verification:
        the drop exists in the current or previous window of its campaign
        (`DROP_NOT_FOUND`, `DROP_EXPIRED`); the fix timestamp is within ±2 minutes
        of the server clock (`STALE_FIX`); distance ≤ collectRadiusMeters +
        min(accuracyMeters, 30) (`TOO_FAR`); the user has not collected the drop
        (`ALREADY_COLLECTED`); the implied speed since the user's last verified
        collect is ≤ 90 m/s (`TELEPORT`), where elapsed time is measured between
        server verification times (client fix timestamps are not trusted for it). A reused `collectId` for the same drop
        and user returns the original receipt (`replayed: true`); for a different
        drop or user it is `409 COLLECT_ID_CONFLICT`. On success a
        `drop.collected` webhook is enqueued. Billable unit `collect`
        (also billed on 422).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CollectRequest' }
      responses:
        '200':
          description: Verified.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CollectResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: collectId already used for a different drop or user (`COLLECT_ID_CONFLICT`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '422':
          description: Verification failed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                tooFar:
                  value: { error: { code: TOO_FAR, message: Fix is 55 m from the drop (allowed 50 m) } }
        '429': { $ref: '#/components/responses/QuotaExceeded' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer dio_...`'
    queryKey:
      type: apiKey
      in: query
      name: key
      description: Only for `GET /v1/worlds/*` and `GET /v1/tiles/*`.
...
  headers:
...
    XDioramaUsage:
      description: '`<used units>/<monthly quota>` after this request.'
      schema: { type: string }
...
    XRateLimitRemaining:
      description: Remaining billable units this month.
      schema: { type: integer }
  responses:
    BadRequest:
      description: '`INVALID_REQUEST` or `INVALID_JSON`.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: '`MISSING_KEY`, `INVALID_KEY`, `MALFORMED_AUTHORIZATION` or `QUERY_KEY_NOT_ALLOWED`.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: '`FORBIDDEN_ROLE`.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    PayloadTooLarge:
      description: '`PAYLOAD_TOO_LARGE` (body > 256 KiB).'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    QuotaExceeded:
      description: '`QUOTA_EXCEEDED` (free plan only).'
      headers:
        X-Diorama-Usage: { $ref: '#/components/headers/XDioramaUsage' }
        X-RateLimit-Remaining: { $ref: '#/components/headers/XRateLimitRemaining' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    ErrorCode:
      type: string
      enum:
        - INVALID_REQUEST
        - INVALID_JSON
        - MISSING_KEY
        - INVALID_KEY
        - MALFORMED_AUTHORIZATION
        - QUERY_KEY_NOT_ALLOWED
        - FORBIDDEN_ROLE
        - NOT_FOUND
        - WEBHOOK_NOT_CONFIGURED
        - COLLECT_ID_CONFLICT
        - PAYLOAD_TOO_LARGE
        - DROP_NOT_FOUND
        - DROP_EXPIRED
        - TOO_FAR
        - STALE_FIX
        - ALREADY_COLLECTED
        - TELEPORT
        - QUOTA_EXCEEDED
        - INTERNAL_ERROR
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: '#/components/schemas/ErrorCode' }
            message: { type: string }
    LngLat:
      type: object
      required: [lng, lat]
      properties:
        lng: { type: number, minimum: -180, maximum: 180 }
        lat: { type: number, minimum: -90, maximum: 90 }
    Bbox:
      type: array
      description: '`[west, south, east, north]`'
      minItems: 4
      maxItems: 4
      items: { type: number }
    DropSpec:
      type: object
      description: '`DropSpec` from `@diorama/protocol`. `id` = `d1.<campaignId>.<window>.<geohash>.<index>`.'
      required: [id, type, coordinate]
      properties:
        id: { type: string }
        type: { type: string, enum: [coin, cd, vinyl, note, model] }
        coordinate: { $ref: '#/components/schemas/LngLat' }
        rarity: { type: string, enum: [common, rare, legendary] }
        value: { type: number }
        payload: {}
...
    DropArea:
      oneOf:
        - type: object
          required: [center, radiusMeters]
          properties:
            center: { $ref: '#/components/schemas/LngLat' }
            radiusMeters: { type: number, minimum: 1, maximum: 50000 }
        - type: object
          required: [bbox]
          properties:
            bbox: { $ref: '#/components/schemas/Bbox' }
    CampaignCreate:
      type: object
      required: [channel, type, rarityWeights, payloadPool, area, density, windowMinutes, startsAt, endsAt, collectRadiusMeters]
      properties:
        channel: { type: string, pattern: '^[A-Za-z0-9_.:-]{1,64}$' }
        type:
          type: string
          enum: [coin, cd, vinyl, note]
          description: '`model` is not supported for campaigns yet.'
        rarityWeights:
          type: object
          properties:
            common: { type: number, minimum: 0 }
            rare: { type: number, minimum: 0 }
            legendary: { type: number, minimum: 0 }
        payloadPool:
          type: array
          minItems: 1
          maxItems: 1000
          items: {}
          description: JSON values; each drop gets one (≤ 64 KiB total).
        area: { $ref: '#/components/schemas/DropArea' }
        density: { type: number, exclusiveMinimum: 0, maximum: 1000, description: Expected drops per km². }
        windowMinutes: { type: integer, minimum: 1, maximum: 1440 }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        collectRadiusMeters: { type: number, minimum: 1, maximum: 500 }
    Campaign:
      allOf:
        - $ref: '#/components/schemas/CampaignCreate'
        - type: object
          required: [id, appId, createdAt]
          properties:
            id: { type: string }
            appId: { type: string }
            createdAt: { type: integer }
    CollectRequest:
      type: object
      required: [dropId, collectId, userId, fix]
      properties:
        dropId: { type: string }
        collectId: { type: string, pattern: '^[A-Za-z0-9_.:-]{8,128}$' }
        userId: { type: string, minLength: 1, maxLength: 128 }
        fix:
          type: object
          required: [lng, lat, accuracyMeters, timestamp]
          properties:
            lng: { type: number }
            lat: { type: number }
            accuracyMeters: { type: number, minimum: 0 }
            timestamp: { type: integer, description: Milliseconds since the Unix epoch. }
    CollectResponse:
      type: object
      required: [receipt, replayed, collect]
      properties:
        receipt:
          type: string
          description: '`base64url(canonical JSON claims).base64url(HMAC-SHA256)`; verify with `verifyReceipt`.'
        replayed: { type: boolean }
        collect:
          type: object
          required: [dropId, collectId, userId, collectedAt]
          properties:
            dropId: { type: string }
            collectId: { type: string }
            userId: { type: string }
            collectedAt: { type: integer }
...